refactor: split ioc and state modules into focused sub-modules (#8)
